Consolidate into a foundational example, remove old natural conversation examples
