+++ examples/foundational/22-filter-incomplete-turns.py
@@ -7,13 +7,17 @@
 import os
 
 from dotenv import load_dotenv
-from idle_handler import IdleHandler
 from loguru import logger
 
 from pipecat.audio.turn.smart_turn.local_smart_turn_v3 import LocalSmartTurnAnalyzerV3
 from pipecat.audio.vad.silero import SileroVADAnalyzer
 from pipecat.audio.vad.vad_analyzer import VADParams
-from pipecat.frames.frames import LLMRunFrame
+from pipecat.frames.frames import (
+    EndTaskFrame,
+    LLMMessagesAppendFrame,
+    LLMRunFrame,
+    TTSSpeakFrame,
+)
 from pipecat.pipeline.pipeline import Pipeline
 from pipecat.pipeline.runner import PipelineRunner
 from pipecat.pipeline.task import PipelineParams, PipelineTask
@@ -22,6 +26,7 @@ from pipecat.processors.aggregators.llm_response_universal import (
     LLMContextAggregatorPair,
     LLMUserAggregatorParams,
 )
+from pipecat.processors.frame_processor import FrameDirection
 from pipecat.runner.types import RunnerArguments
 from pipecat.runner.utils import create_transport
 from pipecat.services.cartesia.tts import CartesiaTTSService
@@ -34,6 +39,81 @@ from pipecat.turns.user_stop import TurnAnalyzerUserTurnStopStrategy
 from pipecat.turns.user_turn_strategies import UserTurnStrategies
 
 load_dotenv(override=True)
+
+
+# System prompts for idle follow-ups
+IDLE_PROMPT_FIRST = """The user has been quiet for a moment. Based on the conversation context, generate a brief, natural follow-up to re-engage them. Your response should:
+- Be contextually relevant to what was just discussed
+- Sound natural and conversational (not robotic)
+- Be concise (1-2 sentences max)
+- Gently prompt them to continue without being pushy
+
+Examples:
+- If you asked a question: "Take your time! I'm curious to hear your thoughts."
+- If discussing a topic: "What do you think about that?"
+- If they seemed engaged: "Are you still there? I'd love to hear more."
+
+Generate ONLY the follow-up message, nothing else."""
+
+IDLE_PROMPT_SECOND = """The user has been quiet for a while now. Generate a brief, friendly check-in message. Your response should:
+- Acknowledge they might be busy or thinking
+- Offer to continue or pause the conversation
+- Be warm and understanding
+- Be very brief (1 sentence)
+
+Examples:
+- "No rush! Let me know if you'd like to continue."
+- "Take your time - I'm here when you're ready."
+- "Should we pick this up later?"
+
+Generate ONLY the check-in message, nothing else."""
+
+
+class IdleHandler:
+    """Helper class to manage user idle retry logic with contextually-aware LLM responses.
+
+    This handler pushes messages through the pipeline to generate contextually appropriate
+    follow-up messages when a user becomes idle. Using the pipeline ensures that
+    interruptions are handled automatically if the user starts speaking.
+    """
+
+    def __init__(self):
+        """Initialize the idle handler."""
+        self._retry_count = 0
+
+    def reset(self):
+        """Reset the retry count when user becomes active."""
+        self._retry_count = 0
+
+    async def handle_idle(self, aggregator):
+        """Handle user idle event with escalating prompts.
+
+        Pushes a system message and triggers LLM through the pipeline, allowing
+        normal interruption handling if the user starts speaking.
+
+        Args:
+            aggregator: The user aggregator that triggered the idle event.
+        """
+        self._retry_count += 1
+        logger.info(f"Handling idle event (attempt {self._retry_count})")
+
+        if self._retry_count <= 2:
+            # Select the appropriate prompt based on retry count
+            prompt = IDLE_PROMPT_FIRST if self._retry_count == 1 else IDLE_PROMPT_SECOND
+
+            # Push through pipeline - this allows interruption handling
+            await aggregator.push_frame(
+                LLMMessagesAppendFrame(messages=[{"role": "system", "content": prompt}])
+            )
+            await aggregator.push_frame(LLMRunFrame())
+        else:
+            # Third attempt: End the conversation gracefully
+            logger.info("User idle timeout reached, ending conversation")
+            await aggregator.push_frame(
+                TTSSpeakFrame("It seems like you're busy right now. Have a nice day!")
+            )
+            await aggregator.push_frame(EndTaskFrame(), FrameDirection.UPSTREAM)
+
 
 # We store functions so objects (e.g. SileroVADAnalyzer) don't get
 # instantiated. The function will be called when the desired transport gets
@@ -109,8 +189,8 @@ async def run_bot(transport: BaseTransport, runner_args: RunnerArguments):
         idle_timeout_secs=runner_args.pipeline_idle_timeout_secs,
     )
 
-    # Initialize idle handler with the LLM for contextual responses
-    idle_handler = IdleHandler(llm)
+    # Initialize idle handler for contextual responses through the pipeline
+    idle_handler = IdleHandler()
 
     @user_aggregator.event_handler("on_user_turn_idle")
     async def on_user_turn_idle(aggregator):
